Доделал: Перенёс функцию в нужный файл, проверка на кнопки самоката и яндекса в отдельном файле, добавил отчёты
