+++ allure_results/0594b26e-cc98-4c90-8dc0-a45f470c8d1b-container.json
@@ -1,0 +1,1 @@
+{"uuid": "7e7a8896-248f-4dc9-8e13-8f3ac8182543", "children": ["75e43726-155e-42d2-9be3-1dc1856dd169"], "befores": [{"name": "driver", "status": "passed", "start": 1724921774323, "stop": 1724921781840}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724921783933, "stop": 1724921787230}], "start": 1724921774323, "stop": 1724921787230}
