+++ allure_results/08d45f02-67c4-4062-9700-e0a47f6e1c85-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920718873, "stop": 1724920718889}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-3')"}], "start": 1724920718873, "stop": 1724920719174}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-3')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-3')"}], "start": 1724920718873, "stop": 1724920719189}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-3')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-3')"}, {"name": "expected_question_text", "value": "'Только начиная с завтрашнего дня. Но скоро станем расторопнее.'"}], "start": 1724920718857, "stop": 1724920719189, "uuid": "12ad6053-c778-46bb-bcbc-0737b983e3a9", "historyId": "b4a1e1f8ee160676af991d428e9427ae", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
