+++ allure_results/0b9b9aa5-df45-4f01-a6aa-590b5f062f92-container.json
@@ -1,0 +1,1 @@
+{"uuid": "528a6665-323e-42cf-be9c-0bd0c4fbba92", "befores": [{"name": "expected_question_text", "status": "passed", "start": 1724920718857, "stop": 1724920718857}], "start": 1724920718857, "stop": 1724920719189}
