+++ allure_results/1800366a-9c4e-4311-861d-d607e1adfc7c-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920748815, "stop": 1724920748846}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-5')"}], "start": 1724920748815, "stop": 1724920749116}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-5')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-5')"}], "start": 1724920748815, "stop": 1724920749131}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-5')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-5')"}, {"name": "expected_question_text", "value": "'Самокат приезжает к вам с полной зарядкой. Этого хватает на восемь суток — даже если будете кататься без передышек и во сне. Зарядка не понадобится.'"}], "start": 1724920748815, "stop": 1724920749131, "uuid": "52d8bedf-066d-4295-97c2-ee2c5ccd0e4b", "historyId": "18c0441df5a8abd16eca425559d884f4", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
