+++ allure_results/21554c61-7e1f-4e0d-a0ac-3d8ce70cea84-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920763580, "stop": 1724920763596}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-6')"}], "start": 1724920763580, "stop": 1724920763875}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-6')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-6')"}], "start": 1724920763580, "stop": 1724920763912}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-6')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-6')"}, {"name": "expected_question_text", "value": "'Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои.'"}], "start": 1724920763580, "stop": 1724920763912, "uuid": "a5f7455e-987a-4542-8594-10dc31c56584", "historyId": "bcd500ebe9fbdec6904881fbee12e8e7", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
