+++ allure_results/51b38b9a-93f8-40cc-bc04-44bbd47a1a10-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка если нажать на логотип Яндекса, в новом окне через редирект откроется главная страница Дзена", "status": "passed", "steps": [{"name": "Лого Яндекса", "status": "passed", "start": 1724921781840, "stop": 1724921781903}], "start": 1724921781840, "stop": 1724921783933, "uuid": "75e43726-155e-42d2-9be3-1dc1856dd169", "historyId": "390eb91731788daca0c0123d7c42dcdb", "testCaseId": "390eb91731788daca0c0123d7c42dcdb", "fullName": "tests.tests_header_button.TestHeaderButton#test_click_yandex_go_to_dzen", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "tests_header_button"}, {"name": "subSuite", "value": "TestHeaderButton"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "1220-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.tests_header_button"}]}
