+++ allure_results/5a4cafef-d00c-45aa-abad-d8c147e77d0e-container.json
@@ -1,0 +1,1 @@
+{"uuid": "0ea68d8a-4789-4486-89d9-26dace3535b6", "children": ["497b778d-0a46-427a-b125-07fe6e36e636"], "befores": [{"name": "driver", "status": "passed", "start": 1724920770989, "stop": 1724920778240}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920778571, "stop": 1724920781844}], "start": 1724920770989, "stop": 1724920781844}
