+++ allure_results/70251d16-3a78-4c13-a9ee-4c5edce9229a-container.json
@@ -1,0 +1,1 @@
+{"uuid": "20e5a4b5-1b3a-44a4-bd9c-b5d8763b3a93", "befores": [{"name": "expected_question_text", "status": "passed", "start": 1724920688890, "stop": 1724920688890}], "start": 1724920688890, "stop": 1724920689206}
