+++ allure_results/788c0398-de45-4cf2-9195-c9147d9063c6-container.json
@@ -1,0 +1,1 @@
+{"uuid": "b27c7294-f5e5-4279-b307-db5fca48fff4", "befores": [{"name": "expected_question_text", "status": "passed", "start": 1724920778240, "stop": 1724920778240}], "start": 1724920778240, "stop": 1724920778571}
