+++ allure_results/7a43a395-ec74-49f9-a088-1d2270ed3756-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920733733, "stop": 1724920733755}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-4')"}], "start": 1724920733733, "stop": 1724920734034}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-4')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-4')"}], "start": 1724920733733, "stop": 1724920734049}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-4')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-4')"}, {"name": "expected_question_text", "value": "'Пока что нет! Но если что-то срочное — всегда можно позвонить в поддержку по красивому номеру 1010.'"}], "start": 1724920733733, "stop": 1724920734049, "uuid": "54197703-d43c-4405-9366-f05c49878139", "historyId": "8294859c0c817ec3e4622289d8a9f40e", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
