+++ allure_results/7ccaf84d-3155-48e5-8cfb-1506f337c354-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920673956, "stop": 1724920673994}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-0')"}], "start": 1724920673956, "stop": 1724920674257}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-0')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-0')"}], "start": 1724920673956, "stop": 1724920674279}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-0')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-0')"}, {"name": "expected_question_text", "value": "'Сутки — 400 рублей. Оплата курьеру — наличными или картой.'"}], "start": 1724920673956, "stop": 1724920674279, "uuid": "b1c8318a-b9c6-445a-befd-f3b79f27008d", "historyId": "ae0df5073bc01d82aedd4534ce9590e1", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
