+++ allure_results/84a7a14c-01e9-427a-bfa4-26e2fa5ab176-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка всего флоу позитивного сценария с двумя наборами данных", "status": "passed", "steps": [{"name": "Верхняя кнопка \"Заказать\"", "status": "passed", "start": 1724921922776, "stop": 1724921923046}, {"name": "Заполнение поля Имя", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921923046, "stop": 1724921923130}, {"name": "Заполнение поля Фамилия", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921923130, "stop": 1724921923161}, {"name": "Заполнение поля Адрес", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921923161, "stop": 1724921923199}, {"name": "Заполнение поля Станция метро", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921923199, "stop": 1724921924365}, {"name": "Заполнение поля Номер телефона", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921924365, "stop": 1724921924434}, {"name": "Клик на кнопку Далее", "status": "passed", "start": 1724921924434, "stop": 1724921924666}, {"name": "Заполнение поля Когда привезти самокат", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921924666, "stop": 1724921925051}, {"name": "Заполнение поля Срок аренды", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921925051, "stop": 1724921926054}, {"name": "Заполнение поля Цвет самоката", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921926054, "stop": 1724921926308}, {"name": "Заполнение поля Комментарий для курьера", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921926308, "stop": 1724921926439}, {"name": "Клик на кнопку Заказать", "status": "passed", "start": 1724921926439, "stop": 1724921926687}, {"name": "Клик на кнопку подтверждения заказа", "status": "passed", "start": 1724921926687, "stop": 1724921926941}, {"name": "Окно Заказ оформлен", "status": "passed", "start": 1724921926941, "stop": 1724921926988}], "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[2]\"), 'first_name': 'Петр', 'last_name': 'Андреев', 'address': 'Москва, ул. Арбат, д. 26', 'metro': 'Преображенская площадь', 'telephone': '89000000002', 'calendar': '30.08.2024', 'rental': ('xpath', \"//div[text()='сутки']\"), 'color': ('xpath', \"//input[@id='grey']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921922776, "stop": 1724921926994, "uuid": "6ba7e042-f586-4104-aa92-a9626c99382b", "historyId": "964b860ae8dde6efee92acc1da113e2e", "testCaseId": "84724b0316555428076238fd3c9ed5b5", "fullName": "tests.test_order_scooter.TestOrder#test_correct_oder", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_order_scooter"}, {"name": "subSuite", "value": "TestOrder"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "11060-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_order_scooter"}]}
