+++ allure_results/85f419a8-e9da-42c2-b627-b856dbbf3ee8-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920688890, "stop": 1724920688927}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-1')"}], "start": 1724920688890, "stop": 1724920689190}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-1')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-1')"}], "start": 1724920688890, "stop": 1724920689206}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-1')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-1')"}, {"name": "expected_question_text", "value": "'Пока что у нас так: один заказ — один самокат. Если хотите покататься с друзьями, можете просто сделать несколько заказов — один за другим.'"}], "start": 1724920688890, "stop": 1724920689206, "uuid": "33f15188-be23-4ef1-a8e6-033fa22d30e6", "historyId": "13865311b38ead75046bacc6bcf51a18", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
