+++ allure_results/8676d0e7-f35b-4bd7-a079-30d3a47ac46d-container.json
@@ -1,0 +1,1 @@
+{"uuid": "12229cdc-b41a-4703-b469-4ce8d71789f4", "children": ["52d8bedf-066d-4295-97c2-ee2c5ccd0e4b"], "befores": [{"name": "driver", "status": "passed", "start": 1724920741405, "stop": 1724920748815}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920749131, "stop": 1724920752247}], "start": 1724920741405, "stop": 1724920752247}
