+++ allure_results/8de5ac5e-809d-4160-bda3-01d0c3b86d28-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920778240, "stop": 1724920778271}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-7')"}], "start": 1724920778240, "stop": 1724920778555}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-7')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-7')"}], "start": 1724920778240, "stop": 1724920778571}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-7')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-7')"}, {"name": "expected_question_text", "value": "'Да, обязательно. Всем самокатов! И Москве, и Московской области.'"}], "start": 1724920778240, "stop": 1724920778571, "uuid": "497b778d-0a46-427a-b125-07fe6e36e636", "historyId": "af2e8e36de28253926b4efd9158aad8f", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
