+++ allure_results/96790b1e-bea7-4ab5-bd45-be05e7296b4a-container.json
@@ -1,0 +1,1 @@
+{"uuid": "907d419b-b1ff-412e-a2c1-0e99143a3618", "children": ["6ba7e042-f586-4104-aa92-a9626c99382b"], "befores": [{"name": "driver", "status": "passed", "start": 1724921915328, "stop": 1724921922776}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724921927010, "stop": 1724921930010}], "start": 1724921915328, "stop": 1724921930010}
