+++ allure_results/99a57160-739f-490b-8f31-2dd99b0c49c7-container.json
@@ -1,0 +1,1 @@
+{"uuid": "b9fc1a51-4134-4859-92a9-ff95be94e158", "children": ["54197703-d43c-4405-9366-f05c49878139"], "befores": [{"name": "driver", "status": "passed", "start": 1724920726317, "stop": 1724920733733}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920734049, "stop": 1724920737377}], "start": 1724920726317, "stop": 1724920737377}
