+++ allure_results/a8e6f6bd-b0ff-405e-b699-54de6c8c113a-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка если нажать на логотип «Самоката», попадёшь на главную страницу «Самоката»", "status": "passed", "steps": [{"name": "Верхняя кнопка \"Заказать\"", "status": "passed", "start": 1724921766755, "stop": 1724921767041}, {"name": "Лого Самокат", "status": "passed", "start": 1724921767041, "stop": 1724921767272}], "start": 1724921766755, "stop": 1724921767272, "uuid": "3c078cbb-f5f5-4767-a9f4-2a01bb9bb564", "historyId": "691bd686f71b56aa1bc818cd49d7475b", "testCaseId": "691bd686f71b56aa1bc818cd49d7475b", "fullName": "tests.tests_header_button.TestHeaderButton#test_correct_go_to_scooter_button", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "tests_header_button"}, {"name": "subSuite", "value": "TestHeaderButton"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "1220-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.tests_header_button"}]}
