+++ allure_results/a975454e-5e9c-4a68-b1ac-d4b458951bb4-container.json
@@ -1,0 +1,1 @@
+{"uuid": "b1eb0642-9078-42c7-b9bc-c4777fb5b4fe", "children": ["bcf0f200-da0a-4636-a423-00d0c49f15f8"], "befores": [{"name": "driver", "status": "passed", "start": 1724921896637, "stop": 1724921903885}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724921908111, "stop": 1724921911302}], "start": 1724921896637, "stop": 1724921911302}
