+++ allure_results/b2368e12-2aee-4b54-8742-c10c1f6c12f0-container.json
@@ -1,0 +1,1 @@
+{"uuid": "849f364c-7900-4dc3-bb25-39eb1e0cd229", "children": ["a5f7455e-987a-4542-8594-10dc31c56584"], "befores": [{"name": "driver", "status": "passed", "start": 1724920756274, "stop": 1724920763574}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920763912, "stop": 1724920766963}], "start": 1724920756274, "stop": 1724920766963}
