+++ allure_results/bbf21eec-a0ec-4139-aee7-fdcf00090b6e-container.json
@@ -1,0 +1,1 @@
+{"uuid": "43abf692-f20b-4bdf-9ac0-5d89c8249c03", "children": ["33f15188-be23-4ef1-a8e6-033fa22d30e6"], "befores": [{"name": "driver", "status": "passed", "start": 1724920681448, "stop": 1724920688890}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920689206, "stop": 1724920692559}], "start": 1724920681448, "stop": 1724920692559}
