+++ allure_results/c19875ba-b769-483a-980f-258a02174d3f-container.json
@@ -1,0 +1,1 @@
+{"uuid": "973099fe-7fd9-4822-a442-c1534c774aa0", "befores": [{"name": "answers_text", "status": "passed", "start": 1724920748815, "stop": 1724920748815}], "start": 1724920748815, "stop": 1724920749131}
