+++ allure_results/c2f6ff7d-4e4c-44cf-9e92-6c78005c2876-container.json
@@ -1,0 +1,1 @@
+{"uuid": "13336ab5-e86a-4252-bc0a-9dde5b95d768", "children": ["3c078cbb-f5f5-4767-a9f4-2a01bb9bb564"], "befores": [{"name": "driver", "status": "passed", "start": 1724921759270, "stop": 1724921766755}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724921767272, "stop": 1724921770295}], "start": 1724921759270, "stop": 1724921770295}
