+++ allure_results/c5a60738-bfc0-4479-b00b-3fb9fae92003-result.json
@@ -1,0 +1,1 @@
+{"name": "Тест проверки текста ответов на вопросы на главной странице веб-приложения", "status": "passed", "steps": [{"name": "Получение текста вопроса", "status": "passed", "steps": [{"name": "Клик на вопрос", "status": "passed", "steps": [{"name": "Переход к списку вопросов", "status": "passed", "start": 1724920703932, "stop": 1724920703963}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-2')"}], "start": 1724920703932, "stop": 1724920704226}], "parameters": [{"name": "question_button_locator", "value": "('id', 'accordion__heading-2')"}, {"name": "question_text_locator", "value": "('id', 'accordion__panel-2')"}], "start": 1724920703932, "stop": 1724920704232}], "parameters": [{"name": "questions", "value": "('id', 'accordion__heading-2')"}, {"name": "answers_text", "value": "('id', 'accordion__panel-2')"}, {"name": "expected_question_text", "value": "'Допустим, вы оформляете заказ на 8 мая. Мы привозим самокат 8 мая в течение дня. Отсчёт времени аренды начинается с момента, когда вы оплатите заказ курьеру. Если мы привезли самокат 8 мая в 20:30, суточная аренда закончится 9 мая в 20:30.'"}], "start": 1724920703932, "stop": 1724920704232, "uuid": "27677483-2027-4951-9367-bdff26384b8c", "historyId": "2cc3498eadcdd9fdecb5c83d8dfc5a55", "testCaseId": "9ce75975fe6f11747333d74223cf3251", "fullName": "tests.test_questions.TestQuestions#test_accordion_questions", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_questions"}, {"name": "subSuite", "value": "TestQuestions"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "10388-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_questions"}]}
