+++ allure_results/c74edacd-e758-4c9e-a21c-e2fe42c8b733-container.json
@@ -1,0 +1,1 @@
+{"uuid": "360f4435-899d-4643-b22d-e175cf3d5b2f", "befores": [{"name": "answers_text", "status": "passed", "start": 1724920673956, "stop": 1724920673956}], "start": 1724920673956, "stop": 1724920674279}
