+++ allure_results/cbec02f6-e33b-4eab-8310-36a50621a32c-container.json
@@ -1,0 +1,1 @@
+{"uuid": "fba6a1ff-c132-4b03-9d3f-94beac750e3a", "children": ["12ad6053-c778-46bb-bcbc-0737b983e3a9"], "befores": [{"name": "driver", "status": "passed", "start": 1724920711334, "stop": 1724920718857}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920719189, "stop": 1724920722289}], "start": 1724920711334, "stop": 1724920722289}
