+++ allure_results/cfd87572-45aa-4ba8-9826-532a57c1c835-container.json
@@ -1,0 +1,1 @@
+{"uuid": "e0ed46ae-4721-488b-9d74-1d3b45c8885d", "befores": [{"name": "user_data", "status": "passed", "start": 1724921922776, "stop": 1724921922776}], "start": 1724921922776, "stop": 1724921926994}
