+++ allure_results/d48ef487-91cf-4d7f-99bf-b44f0333aa37-container.json
@@ -1,0 +1,1 @@
+{"uuid": "a3184319-ed68-423d-bc0d-088fc8df4504", "children": ["b1c8318a-b9c6-445a-befd-f3b79f27008d"], "befores": [{"name": "driver", "status": "passed", "start": 1724920666569, "stop": 1724920673956}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920674279, "stop": 1724920677422}], "start": 1724920666569, "stop": 1724920677422}
