+++ allure_results/d535f1ed-2075-43d4-8b5d-1bc59c77c564-container.json
@@ -1,0 +1,1 @@
+{"uuid": "e8234727-40cf-4f73-aa1c-fe044beb95c5", "befores": [{"name": "user_data", "status": "passed", "start": 1724921903885, "stop": 1724921903885}], "start": 1724921903885, "stop": 1724921908111}
