+++ allure_results/d7e4505a-39ab-4c6c-a99d-57934a440965-container.json
@@ -1,0 +1,1 @@
+{"uuid": "9e230e0b-5e5b-4dcf-91ff-9ff1dc92380a", "children": ["27677483-2027-4951-9367-bdff26384b8c"], "befores": [{"name": "driver", "status": "passed", "start": 1724920696586, "stop": 1724920703932}], "afters": [{"name": "driver::0", "status": "passed", "start": 1724920704248, "stop": 1724920707306}], "start": 1724920696586, "stop": 1724920707306}
