+++ allure_results/e10065a9-2bad-4256-8b76-5c1dde3ae7d9-container.json
@@ -1,0 +1,1 @@
+{"uuid": "5fb365ea-51e3-43f4-a982-df87ef1b50d9", "befores": [{"name": "expected_question_text", "status": "passed", "start": 1724920733733, "stop": 1724920733733}], "start": 1724920733733, "stop": 1724920734049}
