+++ allure_results/ea80bcb2-a88b-4541-8ce6-b5944412da70-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка всего флоу позитивного сценария с двумя наборами данных", "status": "passed", "steps": [{"name": "Верхняя кнопка \"Заказать\"", "status": "passed", "start": 1724921903885, "stop": 1724921904185}, {"name": "Заполнение поля Имя", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921904185, "stop": 1724921904286}, {"name": "Заполнение поля Фамилия", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921904286, "stop": 1724921904317}, {"name": "Заполнение поля Адрес", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921904317, "stop": 1724921904370}, {"name": "Заполнение поля Станция метро", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921904370, "stop": 1724921905504}, {"name": "Заполнение поля Номер телефона", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921905504, "stop": 1724921905573}, {"name": "Клик на кнопку Далее", "status": "passed", "start": 1724921905573, "stop": 1724921905805}, {"name": "Заполнение поля Когда привезти самокат", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921905805, "stop": 1724921906206}, {"name": "Заполнение поля Срок аренды", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921906206, "stop": 1724921907224}, {"name": "Заполнение поля Цвет самоката", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921907224, "stop": 1724921907463}, {"name": "Заполнение поля Комментарий для курьера", "status": "passed", "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921907463, "stop": 1724921907579}, {"name": "Клик на кнопку Заказать", "status": "passed", "start": 1724921907579, "stop": 1724921907811}, {"name": "Клик на кнопку подтверждения заказа", "status": "passed", "start": 1724921907811, "stop": 1724921908065}, {"name": "Окно Заказ оформлен", "status": "passed", "start": 1724921908065, "stop": 1724921908111}], "parameters": [{"name": "user_data", "value": "{'button': ('xpath', \"(//button[contains(@class, 'Button_') and text()='Заказать'])[1]\"), 'first_name': 'Андрей', 'last_name': 'Петров', 'address': 'Москва, ул. Льва Толстого, д. 16', 'metro': 'Преображенская площадь', 'telephone': '89000000001', 'calendar': '29.08.2024', 'rental': ('xpath', \"//div[@class='Dropdown-option' and text()='пятеро суток']\"), 'color': ('xpath', \"//input[@id='black']\"), 'comment': 'Комментарий user_1'}"}], "start": 1724921903885, "stop": 1724921908111, "uuid": "bcf0f200-da0a-4636-a423-00d0c49f15f8", "historyId": "8e6cc583f14c5fc0d95fc18434104536", "testCaseId": "84724b0316555428076238fd3c9ed5b5", "fullName": "tests.test_order_scooter.TestOrder#test_correct_oder", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_order_scooter"}, {"name": "subSuite", "value": "TestOrder"}, {"name": "host", "value": "Home-PC"}, {"name": "thread", "value": "11060-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_order_scooter"}]}
