--- tests/test_go_to_dzen.py
@@ -1,11 +1,0 @@
-from multiprocessing.dummy import current_process
-import pytest
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
-from data import *
-from pages.home_page import HomePageHeader
-from conftest import *
-
-class TestDzen:
-
-    pass
